pyproject.toml
[project]
name = "cat-and-dog"
version = "0.1.0"
description = "A simple web app to classify images of cats and dogs to be debagged and monitored in MLOps."
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "flask>=3.0.3",
    "grafana-api>=1.0.3",
    "gunicorn>=23.0.0",
    "icrawler>=0.6.9",
    "ipykernel>=6.29.5",
    "keras>=3.6.0",
    "mlflow>=2.19.0",
    "numpy>=2.0.2",
    "pillow>=11.0.0",
    "prometheus-client>=0.21.1",
    "prometheus-flask-exporter>=0.23.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.0",
    "tensorflow>=2.18.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.1",
]

